+++ openeogeotrellis/_backend/post_dry_run.py
@@ -7,6 +7,7 @@ import math
 import typing
 from typing import Callable, List, Tuple, Union, Dict, Set, Optional
 
+import pyproj
 from openeo.util import deep_get
 from openeo_driver.backend import AbstractCollectionCatalog, LoadParameters
 from openeo_driver.constants import RESAMPLE_SPATIAL_ALIGN
@@ -352,15 +353,7 @@ def _extract_spatial_extent_from_constraint_load_stac(
         to_resample = extent_orig or extent_aligned
         target_grid = _GridInfo(crs=resample_crs, resolution=resample_resolution)
 
-        # For now, only align the resampling grid with the current spatial extent
-        # (typically the requested extent from `load_collection` or `filter_bbox`)
-        # when the resampling CRS is same as the CRS of the requested extent.
-        # TODO: always align, regardless of involved CRSes? https://github.com/Open-EO/openeo-geopyspark-driver/issues/1662
-        if target_grid.crs_epsg == epsg_code_or_none(to_resample.crs):
-            resample_align = constraint["resample"].get("align")
-        else:
-            resample_align = None
-
+        resample_align = constraint["resample"].get("align")
         extent_resampled = _resample_extent(
             to_resample, crs=resample_crs, resolution=resample_resolution, align=resample_align
         )
@@ -427,14 +420,15 @@ def _resample_extent(
 ) -> BoundingBox:
     target_grid = _GridInfo(crs=crs, resolution=resolution)
 
+    reproject = pyproj.Transformer.from_crs(crs_from=extent.crs, crs_to=target_grid.crs_epsg, always_xy=True).transform
     if align == RESAMPLE_SPATIAL_ALIGN.UPPER_LEFT:
-        offset_x, offset_y = extent.west, extent.north
+        offset_x, offset_y = reproject(extent.west, extent.north)
     elif align == RESAMPLE_SPATIAL_ALIGN.UPPER_RIGHT:
-        offset_x, offset_y = extent.east, extent.north
+        offset_x, offset_y = reproject(extent.east, extent.north)
     elif align == RESAMPLE_SPATIAL_ALIGN.LOWER_LEFT:
-        offset_x, offset_y = extent.west, extent.south
+        offset_x, offset_y = reproject(extent.west, extent.south)
     elif align == RESAMPLE_SPATIAL_ALIGN.LOWER_RIGHT:
-        offset_x, offset_y = extent.east, extent.south
+        offset_x, offset_y = reproject(extent.east, extent.south)
     else:
         if align is not None:
             _log.warning(f"Ignoring invalid resample {align=}")